requirements.txt
fastapi
uvicorn
pymongo
bcrypt
pyjwt
qrcode
python-dotenv